Extract portfolio transaction service boundary
